file names fixed, MSC implemented to register/dev route, checkSessionToken and verifySignature middlewares created
