+++ routes/exams.js
@@ -2,7 +2,7 @@ const express = require("express");
 const Exam = require("../models/Exam");
 const Answer = require("../models/Answer");
 const Question = require("../models/Question");
-const User = require("../models/User");
+const User = require("../models/user.model");
 const ParticipatedUser = require("../models/ParticipatedUser");
 const Score = require("../models/Score");
 const router = express.Router();
